iv_engine/requirements.txt
aiohttp>=3.9.0
numpy>=1.26.0
scipy>=1.12.0
anchorpy>=0.20.0
solders>=0.20.0
solana>=0.34.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
structlog>=24.1.0
asyncio-throttle>=1.0.2
tenacity>=8.2.3
base58>=2.1.1
